requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.0
python-multipart==0.0.6
reedsolo==1.7.0
typer==0.9.0
rich==13.7.0
python-magic==0.4.27
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1